Moved the create-user modal to the create-user template & controller
