+++ app/styles/app.css
@@ -214,6 +214,7 @@ nav a:hover {
   font-size: 16px;
   cursor: pointer;
   box-shadow: 0 2px 5px rgb(0 0 0 / 10%);
+  text-decoration: none;
   transition:
     box-shadow 0.2s ease-in-out,
     background-color 0.3s ease;
@@ -331,7 +332,9 @@ nav a:hover {
   font-size: 16px;
   cursor: pointer;
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-  transition: background-color 0.3s ease, box-shadow 0.2s ease-in-out;
+  transition:
+    background-color 0.3s ease,
+    box-shadow 0.2s ease-in-out;
 }
 
 .submit-button:hover {
