+++ app/templates/users/create-user.hbs
@@ -1,6 +1,22 @@
 {{!-- app/templates/users/create-user.hbs --}}
 {{page-title "Create User"}}
 
-<div class="container">
-  <h1>Create Users</h1>
-</div>
+{{!-- Use reusable Modal for creating a new User --}}
+  <Modal @onClose={{this.closeModal}} @modalHeading="Create a New User">
+    {{!-- Form --}}
+    <form {{on "submit" this.createUser}} class="user-form">
+      <div class="form-group">
+        <label for="userName">User Name:</label>
+        <input id="userName" name="userName" type="text" required class="form-input">
+      </div>
+
+      <div class="form-group">
+        <label for="userDescription">Description:</label>
+        <input id="userDescription" name="userDescription" type="text" required class="form-input">
+      </div>
+
+      <div class="form-actions">
+        <button type="submit" class="submit-button">Create</button>
+      </div>
+    </form>
+  </Modal>
